requirements.txt
streamlit
langchain
openai
faiss-cpu
tiktoken
PyPDF2
python-dotenv
